+++ src/crypto/src/cryptoPoller.hpp
@@ -10,13 +10,13 @@
 class CryptoPoller {
 
   public:
-    CryptoPoller(unsigned int timeoutMilliseconds = 100);
+    CryptoPoller(uint16_t timeoutMilliseconds = 100);
     uint32_t execute(uint32_t instruction);
 
   private:
     static bool isInitialized;
     static uint32_t currentBitFlip;
-    unsigned int timeoutMilliseconds;
+    uint16_t timeoutMilliseconds;
 
     static void writeR0(uint32_t value) { wmem(R0_ADDR, value); }
     static uint32_t readR0() { return rmem(R0_ADDR); }
